Revert "新增优化大厅到LG1-办公层的线路，其他不变"
This reverts commit 3156d7a6fc5fe917ee597b5225c3dd1ace9e4c90.
