+++ src/heist_path/HeistPathA.py
@@ -1,1146 +1,16 @@
 import time
 
 from src.heist_path.HeistPath import HeistPath
-
+from src.heist_path.LobbyGotoLG1Path import LobbyGotoLG1Path
+from src.heist_path.LG1GotoLG2Path import LG1GotoLG2Path
+from src.heist_path.LG2Path import LG2Path
 
 class HeistPathA(HeistPath):
-    # 路径参考自 B站 UP 主：@早柚大魔王
     def run_path(self):
-        self.goto_lg1()
+        LobbyGotoLG1Path(self).run_path()
         self.wait_team_ui_settle()
         self.check_current_floor(1)
-        self.lg1_wp1()
-        self.lg1_wp2()
-        self.lg1_wp3()
-        self.lg1_wp4()
-        idx = self.avoider_strategy_index()
-        if idx == -1:
-            self.lg1_wp5_avoid_combat_01()
-        elif idx == 0:
-            self.lg1_wp5_avoid_combat_02()
-        elif idx == 1:
-            self.lg1_wp5_avoid_combat_03()
+        LG1GotoLG2Path(self).run_path()
         self.wait_team_ui_settle()
         self.check_current_floor(2)
-        self.lg2_wp1_to_exit1()
-        self.lg2_wp1_remains()
-        self.lg2_wp2_to_exit2()
-        self.lg2_wp3_to_layzer_room()
-        self.lg2_wp3_in_layzer_room()
-        self.lg2_wp4()
-        if self.exit_state[1]:
-            self.lg2_wp4_to_exit1()
-        elif self.exit_state[2]:
-            self.lg2_wp4_to_exit2()
-        else:
-            self.lg2_wp4_to_exit3()
-
-
-    def goto_lg1(self):
-        self.log_round_info("寻路到LG1")
-        self.switch_to_runner(check_switched=True)
-        self.sleep(0.81)
-        self.send_key_down("w")
-        self.sleep(0.32)
-        self.send_key_down("lshift")
-        self.sleep(0.16)
-        self.send_key_up("lshift")
-        self.sleep(2.68)
-        self.send_key_down("d")
-        self.sleep(2.55)
-        self.send_key_up("d")
-        self.sleep(0.37)
-        self.wait_and_interact(direction="w", is_lock=True)
-        self.sleep(0.5)
-        self.send_key_down("w")
-        self.sleep(1.8)
-
-        deadline = time.time() + 5.5
-        while time.time() < deadline:
-            self.send_key_down("space")
-            self.sleep(0.06)
-            self.send_key_up("space")
-            self.sleep(0.26)
-
-        if not self.wait_and_interact(direction="w", key_up_sleep=0, is_lock=True):
-            self.goto_lg1_interrupted()
-
-        self.sleep(0.5)
-        self.send_key_down("w")
-        self.sleep(0.2)
-        self.send_key("lshift")
-        self.sleep(0.2)
-        self.send_key_down("d")
-        self.sleep(0.75)
-        self.send_key_up("d")
-        self.sleep(0.01)
-        self.send_key_down("a")
-        self.sleep(1)
-        self.send_key_up("a")
-        self.sleep(0.5)
-        self.send_key_up("w")
-        self.sleep(0.29)
-        self.send_key_down("a")
-        self.sleep(0.29)
-        self.send_key("lshift")
-        self.sleep(3.20)
-        self.send_key_up("a")
-        self.sleep(0.04)
-        self.send_key_down("s")
-        self.sleep(0.29)
-        self.send_key("lshift")
-        self.sleep(1.50)
-        self.send_key_up("s")
-        self.sleep(0.04)
-        self.send_key_down("d")
-        self.sleep(0.29)
-        self.send_key("lshift")
-        self.sleep(2.50)
-        self.send_key_up("d")
-        self.sleep(0.40)
-        self.send_key_down("a")
-        self.sleep(0.71)
-        self.send_key_up("a")
-        self.sleep(0.36)
-        self.send_key_down("s")
-        self.sleep(1.50)
-        self.send_key_up("s")
-        self.sleep(0.14)
-        self.send_key_down("f") # start pick
-        self.sleep(0.04)
-        self.send_key_down("w")
-        self.sleep(2.5)
-        self.send_key_up("w")
-        self.sleep(0.10)
-        self.send_key_up("f") # end pick
-        self.sleep(0.13)
-        self.send_key_down("s")
-        self.sleep(0.14)
-        self.send_key_up("s")
-        self.sleep(0.20)
-        self.clear_current_combat()
-        self.send_key_down("f")
-        self.sleep(0.5)
-        self.send_key_down("w")
-        self.sleep(0.22)
-        self.send_key_down("lshift")
-        self.sleep(0.10)
-        self.send_key_up("lshift")
-        self.sleep(1)
-        self.send_key_up("f")
-        self.sleep(0.1)
-        self.send_key_down("d")
-        self.sleep(1)
-        self.send_key_up("w")
-        self.sleep(1.5)
-        self.send_key_up("d")
-        self.sleep(0.35)
-        self.send_key_down("a")
-        self.sleep(0.88)
-        self.send_key_up("a")
-        self.sleep(0.30)
-        self.send_key_down("s")
-        self.sleep(0.43)
-        self.send_key_down("lshift")
-        self.sleep(0.13)
-        self.send_key_up("lshift")
-        self.sleep(1.4)
-        self.send_key_down("a")
-        self.sleep(0.53)
-        self.send_key_up("a")
-        self.sleep(1.64)
-        self.wait_and_interact(direction="s")
-        self.sleep(0.50)
-        self.send_key_down("s")
-        self.sleep(0.10)
-        self.wait_and_interact(direction="s")
-
-    def goto_lg1_interrupted(self):
-        self.clear_current_combat()
-        self.send_key_down("w")
-        self.sleep(2.02)
-        self.send_key_up("w")
-        self.sleep(0.51)
-        self.send_key_down("s")
-        self.sleep(0.60)
-        self.send_key_up("s")
-        self.sleep(0.22)
-        self.send_key_down("d")
-        self.sleep(3.41)
-        self.send_key_up("d")
-        self.sleep(0.32)
-        self.send_key_down("a")
-        self.sleep(1.16)
-        self.send_key_up("a")
-        self.sleep(0.20)
-        self.send_key_down("w")
-        self.sleep(1.51)
-        self.send_key_up("w")
-        self.sleep(0.11)
-        self.wait_and_interact(direction="w", is_lock=True)
-        self.sleep(0.5)
-
-    def lg1_wp1(self):
-        self.sleep(0.75)
-        self.send_key_down("w")
-        self.sleep(9.06)
-        self.send_key_up("w")
-        self.sleep(0.51)
-        self.send_key_down("d")
-        self.sleep(1.71)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_down("s")
-        self.sleep(1.01)
-        self.wait_and_interact(direction="s")
-        self.sleep(0.11)
-        self.send_key_down("f")  # start pick
-        self.sleep(0.78)
-        self.send_key_down("s")
-        self.sleep(1.25)
-        self.send_key_up("s")
-        self.sleep(0.11)
-        self.send_key_down("d")
-        self.sleep(3.03)
-        self.send_key_up("d")
-        self.sleep(0.22)
-        self.send_key_down("a")
-        self.sleep(3.90)
-        self.send_key_up("a")
-        self.sleep(0.31)
-        self.send_key_down("d")
-        self.sleep(0.40)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(2.01)
-        self.send_key_up("w")
-        self.sleep(0.11)
-        self.send_key_down("d")
-        self.sleep(5.60)
-        self.send_key_up("d")
-        self.sleep(0.06)
-        self.send_key_down("w")
-        self.sleep(2.02)
-        self.send_key_up("w")
-        self.sleep(0.11)
-        self.send_key_down("d")
-        self.sleep(3.21)
-        self.send_key_up("d")
-        self.sleep(0.12)
-
-    def lg1_wp2(self):
-        self.send_key_down("d")
-        self.sleep(1.80)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(0.71)
-        self.send_key_up("w")
-        self.sleep(0.11)
-        self.send_key_down("a")
-        self.sleep(0.71)
-        self.send_key_up("a")
-        self.sleep(0.11)
-        self.send_key_down("d")
-        self.sleep(2.28)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(1.26)
-        self.send_key_up("w")
-        self.sleep(0.11)
-        self.send_key_up("f")  # end pick
-        self.sleep(2.93)
-        self.send_key_down("w")  # 过镭射1
-        self.sleep(2.01)
-        self.send_key_up("w")
-        self.sleep(0.44)
-        self.start_interaction_watch()
-        self.send_key_down("w")  # 过镭射2
-        self.sleep(8.51)
-        self.send_key_up("w")
-        self.stop_interaction_watch()
-        self.sleep(0.33)
-
-    def lg1_wp3(self):
-        self.send_key_down("f")  # start pick
-        self.sleep(0.11)
-        self.send_key_down("a")
-        self.sleep(2.13)
-        self.send_key_up("a")
-        self.sleep(0.52)
-        self.send_key_down("s")
-        self.sleep(1.32)
-        self.send_key_up("s")
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(0.20)
-        self.send_key_up("w")
-        self.sleep(0.31)
-        self.send_key_down("a")
-        self.sleep(1.50)
-        self.send_key_up("a")
-        self.sleep(0.11)
-        self.send_key_down("d")
-        self.sleep(0.20)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.start_interaction_watch()
-        self.send_key_down("w")
-        self.sleep(3.19)
-        self.send_key_down("a")
-        self.sleep(0.31)
-        self.send_key_up("a")
-        self.sleep(5.16)
-        self.send_key_up("w")
-        self.stop_interaction_watch()
-        self.sleep(0.11)
-        self.send_key_down("s")
-        self.sleep(0.15)
-        self.send_key_up("s")
-        self.sleep(0.11)
-        self.send_key_up("f")  # end pick
-        self.sleep(0.11)
-        self.send_key_down("d")
-        self.sleep(2.51)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_down("a")
-        self.sleep(0.40)
-        self.send_key_up("a")
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(5.31)
-        self.send_key_up("w")
-        self.sleep(0.12)
-
-    def lg1_wp4(self):
-        self.send_key_down("d")
-        self.sleep(0.11)
-        self.send_key_down("s")
-        self.sleep(3.31)
-        self.send_key_up("s")
-        self.sleep(0.12)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_down("a")
-        self.sleep(0.31)
-        self.send_key_up("a")
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(1.50)
-        self.send_key_up("w")
-        self.sleep(0.11)
-        self.start_interaction_watch()
-        self.send_key_down("f")  # start pick
-        self.sleep(0.11)
-        self.send_key_down("a")
-        self.sleep(0.11)
-        self.send_key_up("a")
-        self.sleep(1.22)
-        self.stop_interaction_watch()
-        self.send_key_down("w")
-        self.sleep(6.58)
-        self.send_key_down("d")
-        self.sleep(2.62)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_up("w")
-        self.sleep(0.11)
-        self.send_key_down("a")
-        self.sleep(0.31)
-        self.send_key_up("a")
-        self.sleep(0.32)
-        self.send_key_down("w")
-        self.sleep(0.21)
-        self.send_key_down("space")
-        self.sleep(0.06)
-        self.send_key_up("space")
-        self.sleep(0.31)
-        self.send_key_down("space")
-        self.sleep(0.06)
-        self.send_key_up("space")
-        self.sleep(1.30)
-        self.start_interaction_watch()
-        self.send_key_down("d")
-        self.sleep(2.10)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_up("w")
-        self.sleep(0.65)
-        self.send_key_down("w")
-        self.sleep(0.22)
-        self.send_key_down("d")
-        self.sleep(0.61)
-        self.send_key_down("space")
-        self.sleep(0.06)
-        self.send_key_up("space")
-        self.sleep(0.56)
-        self.send_key_down("space")
-        self.sleep(0.06)
-        self.send_key_up("space")
-        self.sleep(0.14)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_up("w")
-        self.sleep(0.34)
-        self.send_key_down("d")
-        self.sleep(1.41)
-        self.stop_interaction_watch()
-        self.send_key_down("w")
-        self.sleep(0.81)
-        self.send_key_up("w")
-        self.sleep(0.11)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_down("s")
-        self.sleep(0.47)
-        self.send_key_up("s")
-        self.sleep(0.11)
-        self.send_key_down("d")
-        self.sleep(0.60)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(3.38)
-        self.send_key_up("w")
-        self.sleep(0.34)
-        self.send_key_down("d")
-        self.sleep(0.61)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.loot_safes_while_walking(direction="s", time_out=2.37)
-        self.sleep(0.11)
-        self.send_key_up("f")  # end pick
-        self.sleep(0.10)
-        self.send_key_down("d")
-        self.sleep(1.33)
-        self.send_key_up("d")
-        self.sleep(0.12)
-        self.send_key_down("w")
-        self.sleep(9.40)
-        self.send_key_up("w")
-        self.sleep(0.31)
-
-    def lg1_wp5_avoid_combat_01(self):
-        self.send_key_down("w")
-        self.sleep(2.02)
-        self.send_key_up("w")
-        self.sleep(0.10)
-        self.send_key_down("s")
-        self.sleep(0.11)
-        self.send_key_down("lshift")
-        self.sleep(0.06)
-        self.send_key_up("lshift")
-        self.sleep(0.81)
-        self.send_key_up("s")
-        self.sleep(2.01)
-        self.send_key_down("w")
-        self.sleep(0.11)
-
-        deadline = time.time() + 4.5
-        while time.time() < deadline:
-            self.send_key("lshift")
-            self.sleep(0.51)
-
-        self.wait_and_interact(direction="w", is_lock=True)
-        self.sleep(0.11)
-        self.switch_to_runner(check_switched=True)
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(1.00)
-        self.wait_and_interact(direction="w")
-
-    def lg1_wp5_avoid_combat_02(self):
-        self.send_key_down("s")
-        self.sleep(1.50)
-        self.send_key_up("s")
-        self.sleep(0.11)
-
-        self.switch_to_avoider(check_switched=True)
-        self.sleep(0.5)
-        self.perform_avoidance_action()
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(6.0)
-        self.send_key_up("w")
-        self.switch_to_runner(check_switched=True)
-        self.sleep(0.5)
-        self.wait_and_interact(is_lock=True)
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(1.00)
-        self.wait_and_interact(direction="w")
-
-    def lg1_wp5_avoid_combat_03(self):
-        self.switch_to_avoider(check_switched=True)
-        self.sleep(0.5)
-        self.perform_avoidance_action()
-        self.sleep(3.2)
-        self.send_key_down("w")
-        self.sleep(0.11)
-
-        deadline = time.time() + 4.5
-        while time.time() < deadline:
-            self.send_key("lshift")
-            self.sleep(0.51)
-
-        self.send_key_up("w")
-        self.sleep(0.2)
-        self.perform_avoidance_action()
-        self.sleep(3.2)
-        self.switch_to_runner(check_switched=True)
-        self.sleep(0.5)
-        self.wait_and_interact(is_lock=True)
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(1.00)
-        self.wait_and_interact(direction="w")
-
-    def lg2_wp1_to_exit1(self):
-        self.sleep(1.25)
-        self.send_key_down("w")
-        self.sleep(5.04)
-        self.send_key_up("w")
-        self.sleep(0.13)
-        self.send_key_down("f") # start pick
-        self.sleep(0.11)
-        self.send_key_down("a")
-        self.sleep(8.07)
-        self.send_key_up("a")
-        self.sleep(0.11)
-        self.send_key_down("s")
-        self.sleep(0.11)
-        self.send_key_down("d")
-        self.sleep(1.51)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_up("s")
-        self.sleep(0.33)
-        self.send_key_down("w")
-        self.sleep(0.51)
-        self.send_key_down("d")
-        self.sleep(1.41)
-        self.send_key_up("d")
-        self.sleep(1.21)
-        self.send_key_up("w")
-        self.sleep(0.11)
-        self.send_key_down("d")
-        self.sleep(0.72)
-        self.send_key_up("d")
-        self.sleep(0.59)
-        self.send_key_down("s")
-        self.sleep(0.51)
-        self.send_key_up("s")
-        self.sleep(0.11)
-        self.send_key_down("d")
-        self.sleep(0.89)
-        self.send_key_up("d")
-        self.sleep(0.41)
-        self.send_key_up("f") # end pick
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(2.81)
-        self.exit_state[1] = self.try_open_exit(direction="w")
-        self.sleep(0.40)
-    
-    def lg2_wp1_remains(self):
-        self.send_key_down("w")
-        self.sleep(2.10)
-        self.send_key_up("w")
-        self.sleep(0.11)
-        self.send_key_down("f") # start pick
-        self.sleep(0.11)
-        self.send_key_down("a")
-        self.sleep(0.90)
-        self.send_key_up("a")
-        self.sleep(0.30)
-        self.send_key_down("w")
-        self.sleep(0.80)
-        self.send_key_up("w")
-        self.sleep(0.11)
-        self.send_key_down("d")
-        self.sleep(0.62)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(3.06)
-        self.send_key_up("w")
-        self.sleep(0.11)
-        self.switch_to_runner()
-        self.sleep(0.11)
-        self.send_key_down("a")
-        self.sleep(0.81)
-        self.send_key_up("a")
-        self.sleep(0.30)
-        self.send_key_down("d")
-        self.sleep(2.73)
-        self.send_key_up("d")
-        self.switch_to_runner()
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(1.01)
-        self.send_key_up("w")
-        self.sleep(0.18)
-        self.send_key_down("d")
-        self.sleep(0.70)
-        self.send_key_up("d")
-        self.sleep(0.12)
-        self.send_key_down("s")
-        self.sleep(3.02)
-        self.send_key_up("s")
-        self.switch_to_runner()
-        self.sleep(0.72)
-        self.send_key_down("s")
-        self.sleep(6.36)
-        self.send_key_up("s")
-        self.switch_to_runner()
-        self.sleep(0.11)
-        self.send_key_down("d")
-        self.sleep(4.56)
-        self.send_key_up("d")
-        self.sleep(1.23)
-        self.loot_safes_while_walking(direction="w", time_out=3.06, hold=True)
-        self.switch_to_runner()
-        self.send_key_down("d")
-        self.sleep(0.73)
-        self.send_key_up("d")
-        self.sleep(6.16)
-        self.send_key_down("d")
-        self.sleep(1.31)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_up("w")
-        self.switch_to_runner()
-        self.sleep(0.14)
-        self.send_key_down("s")
-        self.sleep(0.22)
-        self.send_key_up("s")
-        self.sleep(0.11)
-        self.send_key_down("d")
-        self.sleep(0.81)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(0.71)
-        self.send_key_up("w")
-        self.sleep(0.30)
-        self.send_key_down("d")
-        self.sleep(0.72)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_down("s")
-        self.sleep(0.90)
-        self.send_key_up("s")
-        self.sleep(0.11)
-        self.send_key_down("d")
-        self.sleep(2.02)
-        self.send_key_up("d")
-        self.switch_to_runner()
-        self.sleep(0.30)
-        self.send_key_down("a")
-        self.sleep(0.60)
-        self.send_key_up("a")
-        self.sleep(0.11)
-        self.send_key_down("s")
-        self.sleep(3.26)
-        self.send_key_up("s")
-        self.switch_to_runner()
-        self.sleep(0.11)
-        self.send_key_down("d")
-        self.sleep(1.01)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_down("s")
-        self.sleep(0.51)
-        self.send_key_up("s")
-        self.sleep(0.29)
-        self.send_key_down("a")
-        self.sleep(0.61)
-        self.send_key_up("a")
-        self.sleep(0.11)
-        self.send_key_down("s")
-        self.sleep(7.12)
-        self.send_key_up("s")
-        self.switch_to_runner()
-        self.sleep(0.12)
-        self.send_key_down("w")
-        self.sleep(0.51)
-        self.send_key_down("d")
-        self.sleep(1.44)
-        self.send_key_up("d")
-        self.sleep(0.91)
-        self.send_key_up("w")
-        self.sleep(0.30)
-        self.send_key_down("d")
-        self.sleep(0.72)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_down("s")
-        self.sleep(0.61)
-        self.send_key_up("s")
-        self.sleep(0.11)
-        self.send_key_up("f") # end pick
-        self.sleep(0.11)
-
-    def lg2_wp2_to_exit2(self):
-        self.send_key_down("f") # start pick
-        self.sleep(0.11)
-        self.send_key_down("d")
-        self.sleep(1.21)
-        self.send_key_down("space")
-        self.sleep(0.06)
-        self.send_key_up("space")
-        self.sleep(0.81)
-        self.send_key_up("d")
-        self.sleep(0.12)
-        self.send_key_down("w")
-        self.sleep(1.70)
-        self.send_key_up("w")
-        self.sleep(0.11)
-        self.send_key_up("f") # end pick
-        self.sleep(0.11)
-        self.send_key_down("d")
-        self.sleep(4.52)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_down("s")
-        self.sleep(0.31)
-        self.send_key_up("s")
-        self.sleep(0.11)
-        self.send_key_down("d")
-        self.sleep(0.81)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(3.96)
-        self.send_key_up("w")
-        self.sleep(0.11)
-        self.send_key_down("f") # start pick
-        self.sleep(0.15)
-        self.send_key_down("a")
-        self.sleep(0.71)
-        self.send_key_up("a")
-        self.sleep(0.31)
-        self.send_key_down("d")
-        self.sleep(1.61)
-        self.send_key_up("d")
-        self.sleep(0.30)
-        self.send_key_down("a")
-        self.sleep(0.72)
-        self.send_key_up("a")
-        self.sleep(0.26)
-        self.send_key_down("w")
-        self.sleep(2.47)
-        self.send_key_up("w")
-        self.sleep(0.11)
-        self.send_key_down("d")
-        self.sleep(2.31)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(4.03)
-        self.send_key_up("w")
-        self.sleep(0.11)
-        self.send_key_down("s")
-        self.sleep(2.87)
-        self.send_key_up("s")
-        self.sleep(0.11)
-        self.send_key_down("d")
-        self.sleep(1.51)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_up("f") # end pick
-        self.sleep(0.11)
-        self.send_key_down("s")
-        self.sleep(0.60)
-        self.send_key_up("s")
-        self.sleep(0.28)
-        self.send_key_down("w")
-        self.sleep(2.56)
-        self.send_key_down("a")
-        self.sleep(0.40)
-        self.send_key_up("a")
-        self.sleep(1.57)
-        self.exit_state[2] = self.try_open_exit(direction="w")
-        self.sleep(0.40)
-        self.switch_to_runner()
-        self.sleep(0.5)
-
-    def lg2_wp3_to_layzer_room(self):
-        self.send_key_down("a")
-        self.sleep(3.03)
-        self.send_key_down("f") # start pick
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(2.55)
-        self.send_key_up("w")
-        self.sleep(0.51)
-        self.send_key_up("a")
-        self.sleep(0.11)
-        self.send_key_down("s")
-        self.sleep(0.56)
-        self.send_key_up("s")
-        self.sleep(1.18)
-        self.send_key_down("a")
-        self.sleep(2.61)
-        self.send_key_up("a")
-        self.sleep(0.11)
-        self.send_key_down("s")
-        self.sleep(1.77)
-        self.send_key_up("s")
-        self.sleep(0.11)
-        self.send_key_down("a")
-        self.sleep(0.60)
-        self.send_key_up("a")
-        self.sleep(0.29)
-        self.send_key_down("d")
-        self.sleep(1.31)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_down("s")
-        self.sleep(0.76)
-        self.send_key_up("s")
-        self.sleep(0.30)
-        self.send_key_down("a")
-        self.sleep(0.61)
-        self.send_key_up("a")
-        self.sleep(0.11)
-        self.send_key_up("f") # end pick
-        self.sleep(0.11)
-        self.send_key_down("s")
-        self.sleep(2.96)
-        self.send_key_up("s")
-
-    def lg2_wp3_in_layzer_room(self):
-        self.send_key_down("d")
-        self.sleep(0.36)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_down("f") # start pick
-        self.sleep(0.11)
-        self.wait_for_safe_loot(time_out=0.8)
-        self.sleep(0.11)
-        self.send_key_down("a")
-        self.sleep(0.46)
-        self.send_key_up("a")
-
-        self.sleep(0.20)
-        self.send_key_down("w")
-        self.sleep(0.40)
-        self.send_key_up("w")
-        self.sleep(0.36)
-        self.send_key_down("a")
-        self.sleep(0.51)
-        self.send_key_down("w")
-        self.sleep(1.00)
-        self.send_key_up("a")
-        self.send_key_up("w")
-
-        self.send_key_down("d")
-        self.sleep(0.60)
-        self.send_key_up("d")
-        self.sleep(0.30)
-        self.send_key_down("s")
-        self.sleep(0.30)
-        self.send_key_up("s")
-
-        self.sleep(0.11)
-        self.send_key_down("a")
-        self.sleep(0.11)
-        self.send_key_down("s")
-        self.sleep(1.41)
-        self.send_key_up("a")
-        self.sleep(0.11)
-        self.send_key_up("s")
-        self.sleep(0.11)
-        self.send_key_down("d")
-        self.sleep(1.30)
-        self.send_key_up("d")
-        self.sleep(0.54)
-        self.send_key_down("a")
-        self.sleep(0.31)
-        self.send_key_up("a")
-        self.sleep(0.11)
-        self.send_key_down("s")
-        self.sleep(0.31)
-        self.send_key_up("s")
-        self.sleep(0.16)
-        self.send_key_down("a")
-        self.sleep(0.33)
-        self.send_key_down("space")
-        self.sleep(0.06)
-        self.send_key_up("space")
-        self.sleep(0.41)
-        self.send_key_down("space")
-        self.sleep(0.06)
-        self.send_key_up("space")
-        self.sleep(0.31)
-        self.send_key_up("a")
-        self.sleep(0.40)
-        self.send_key_down("s")
-        self.sleep(0.20)
-        self.send_key_down("space")
-        self.sleep(0.07)
-        self.send_key_up("space")
-        self.sleep(1.21)
-        self.send_key_up("s")
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(1.31)
-        self.send_key_up("w")
-        self.sleep(0.11)
-        self.send_key_down("a")
-        self.sleep(0.74)
-        self.send_key_up("a")
-        self.sleep(0.80)
-        self.send_key_down("d")
-        self.sleep(0.71)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_up("f") # end pick
-        self.sleep(0.11)
-        self.send_key_down("s")
-        self.sleep(0.31)
-        self.send_key_down("d")
-        self.sleep(1.52)
-        self.send_key_up("s")
-        self.sleep(0.61)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_down("f") # start pick
-        self.wait_for_safe_loot(raise_timeout=True)
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(0.11)
-        self.send_key_down("d")
-        self.sleep(0.90)
-        self.send_key_up("d")
-        self.sleep(1.32)
-        self.send_key_up("w")
-        self.sleep(0.92)
-        self.send_key_down("s")
-        self.sleep(0.08)
-        self.send_key_down("d")
-        self.sleep(1.36)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_up("s")
-        self.sleep(0.11)
-        self.send_key_down("a")
-        self.sleep(0.51)
-        self.send_key_up("a")
-        self.sleep(0.11)
-        self.send_key_down("s")
-        self.sleep(1.17)
-        self.send_key_up("s")
-        self.sleep(0.10)
-        self.send_key_down("a")
-        self.sleep(2.40)
-        self.send_key_up("a")
-        self.sleep(0.11)
-        self.send_key_down("s")
-        self.sleep(0.63)
-        self.send_key_up("s")
-        self.sleep(0.11)
-        self.send_key_down("a")
-        self.sleep(0.51)
-        self.send_key_up("a")
-        self.sleep(0.13)
-        self.send_key_down("s")
-        self.sleep(0.11)
-        self.send_key_down("space")
-        self.sleep(0.06)
-        self.send_key_up("space")
-        self.sleep(0.47)
-        self.send_key_down("space")
-        self.sleep(0.06)
-        self.send_key_up("space")
-        self.sleep(0.51)
-        self.send_key_up("s")
-        self.sleep(0.12)
-        self.send_key_down("a")
-        self.sleep(3.01)
-        self.send_key_up("a")
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(1.51)
-        self.send_key_up("w")
-        self.sleep(0.13)
-        self.send_key_down("d")
-        self.sleep(0.11)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(0.31)
-        self.send_key_down("space")
-        self.sleep(0.06)
-        self.send_key_up("space")
-        self.sleep(0.81)
-        self.send_key_down("space")
-        self.sleep(0.06)
-        self.send_key_up("space")
-        self.sleep(0.51)
-        self.send_key_down("space")
-        self.sleep(0.06)
-        self.send_key_up("space")
-        self.sleep(0.51)
-        self.send_key_down("d")
-        self.sleep(0.51)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_up("w")
-        self.sleep(0.22)
-        self.send_key_down("s")
-        self.sleep(0.03)
-        self.send_key_down("space")
-        self.sleep(0.06)
-        self.send_key_up("space")
-        self.sleep(0.22)
-        self.send_key_down("space")
-        self.sleep(0.06)
-        self.send_key_up("space")
-        self.sleep(0.51)
-        self.send_key_up("s")
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(0.11)
-        self.send_key_down("a")
-        self.sleep(1.21)
-        self.send_key_up("a")
-        self.sleep(0.11)
-        self.send_key_up("w")
-        self.sleep(0.22)
-        self.send_key_down("d")
-        self.sleep(0.60)
-        self.send_key_up("d")
-        self.sleep(0.31)
-        self.send_key_down("s")
-        self.sleep(0.40)
-        self.send_key_up("s")
-        self.sleep(0.12)
-        self.send_key_down("a")
-        self.sleep(1.31)
-        self.send_key_up("a")
-        self.sleep(0.11)
-
-    def lg2_wp4(self):
-        self.send_key_down("w")
-        self.sleep(5.53)
-        self.send_key_up("w")
-        self.sleep(0.11)
-        self.send_key_down("s")
-        self.sleep(0.40)
-        self.send_key_up("s")
-        self.sleep(0.11)
-        self.send_key_down("a")
-        self.sleep(2.91)
-        self.send_key_up("a")
-        self.sleep(0.11)
-        self.send_key_up("f") # end pick
-        self.sleep(0.12)
-
-    def lg2_wp4_to_exit1(self):
-        self.send_key_down("a")
-        self.sleep(0.11)
-        self.send_key_down("s")
-        self.sleep(0.71)
-        self.send_key_up("s")
-        self.sleep(6.51)
-        self.send_key_up("a")
-        self.sleep(0.29)
-        self.send_key_down("d")
-        self.sleep(0.31)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_down("f") # start pick
-        self.sleep(0.11)
-        self.send_key_down("s")
-        self.sleep(0.11)
-        self.send_key_down("lshift")
-        self.sleep(0.06)
-        self.send_key_up("lshift")
-        self.sleep(0.11)
-        self.send_key_down("lshift")
-        self.sleep(0.06)
-        self.send_key_up("lshift")
-        self.sleep(4.90)
-        self.send_key_up("s")
-        self.sleep(0.11)
-        self.send_key_down("s")
-        self.sleep(0.11)
-        self.send_key_up("f") # end pick
-        self.sleep(0.41)
-        self.send_key_down("space")
-        self.sleep(0.06)
-        self.send_key_up("space")
-        self.sleep(0.56)
-        self.send_key_down("space")
-        self.sleep(0.06)
-        self.send_key_up("space")
-        self.sleep(0.55)
-        self.send_key_down("space")
-        self.sleep(0.06)
-        self.send_key_up("space")
-        self.sleep(1.21)
-        self.send_key_up("s")
-        self.sleep(0.11)
-        self.send_key_down("a")
-        self.sleep(0.31)
-        self.send_key_up("a")
-        self.sleep(0.11)
-        self.walk_until_extract_panel(direction="w")
-
-    def lg2_wp4_to_exit2(self):
-        self.send_key_down("w")
-        self.sleep(0.11)
-        self.send_key_down("lshift")
-        self.sleep(0.06)
-        self.send_key_up("lshift")
-        self.sleep(0.11)
-        self.send_key_down("lshift")
-        self.sleep(0.06)
-        self.send_key_up("lshift")
-        self.sleep(1.10)
-        self.send_key_down("d")
-        self.sleep(1.21)
-        self.send_key_up("w")
-        self.sleep(1.50)
-        self.send_key_down("w")
-        self.sleep(1.01)
-        self.send_key_up("w")
-        self.sleep(4.81)
-        self.send_key_up("d")
-        self.sleep(0.41)
-        self.send_key_down("s")
-        self.sleep(1.00)
-        self.send_key_up("s")
-        self.sleep(0.11)
-        self.walk_until_extract_panel(direction="d")
-
-    def lg2_wp4_to_exit3(self):
-        self.send_key_down("w")
-        self.sleep(0.14)
-        self.send_key_down("lshift")
-        self.sleep(0.13)
-        self.send_key_up("lshift")
-        self.sleep(2.70)
-        self.send_key_down("a")
-        self.sleep(1.98)
-        self.send_key_up("a")
-        self.wait_and_interact(direction="w", is_lock=True, time_out=6)
-        self.sleep(0.20)
-        self.send_key_down("w")
-        self.sleep(0.05)
-        self.send_key_down("lshift")
-        self.sleep(0.22)
-        self.send_key_down("d")
-        self.sleep(0.05)
-        self.send_key_up("lshift")
-        self.sleep(1.76)
-        self.send_key_up("w")
-        self.sleep(0.26)
-        self.send_key_up("d")
-        self.sleep(0.10)
-        self.walk_until_extract_panel(direction="d")
+        LG2Path(self).run_path()
