--- src/heist_path/HeistPathB.py
@@ -1,151 +1,0 @@
-import time
-import re
-
-from src.heist_path.HeistPath import HeistPath
-from src.heist_path.HeistPathA import HeistPathA
-
-# 优化了大厅到LG1，其他目前不变
-class LobbyGotoLG1Path(HeistPath):
-    def run_path(self):
-        self.goto_lg1_skip()
-        self.wait_team_ui_settle()
-        self.check_current_floor(1)
-        self.switch_to_runner(check_switched=True)
-        HeistPathA(self).lg1_wp1()
-        HeistPathA(self).lg1_wp2()
-        HeistPathA(self).lg1_wp3()
-        HeistPathA(self).lg1_wp4()
-        idx = self.avoider_strategy_index()
-        if idx == -1:
-            HeistPathA(self).lg1_wp5_avoid_combat_01()
-        elif idx == 0:
-            HeistPathA(self).lg1_wp5_avoid_combat_02()
-        elif idx == 1:
-            HeistPathA(self).lg1_wp5_avoid_combat_03()
-        self.wait_team_ui_settle()
-        self.check_current_floor(2)
-        HeistPathA(self).lg2_wp1_to_exit1()
-        HeistPathA(self).lg2_wp1_remains()
-        HeistPathA(self).lg2_wp2_to_exit2()
-        HeistPathA(self).lg2_wp3_to_layzer_room()
-        HeistPathA(self).lg2_wp3_in_layzer_room()
-        HeistPathA(self).lg2_wp4()
-        if self.exit_state[1]:
-            HeistPathA(self).lg2_wp4_to_exit1()
-        elif self.exit_state[2]:
-            HeistPathA(self).lg2_wp4_to_exit2()
-        else:
-            HeistPathA(self).lg2_wp4_to_exit3()
-    
-    # 大厅到LG1，修改版
-    def goto_lg1_skip(self):
-        self.log_round_info("寻路到LG1")
-        self.sleep(0.30)
-        self.switch_to_runner(check_switched=True)
-        self.sleep(0.10)
-        self.send_key_down('w')
-        self.sleep(0.42)
-        self.send_key('lshift', down_time=0.25)
-        self.sleep(0.42)
-        self.send_key('lshift', down_time=0.25)
-        self.sleep(0.42)
-        self.send_key('lshift', down_time=0.25)
-        self.sleep(0.42)
-        self.send_key('lshift', down_time=0.25)
-        self.sleep(0.42)
-        self.send_key_up('w')
-        self.sleep(0.20)
-        self.send_key_down('d')
-        self.sleep(0.57)
-        self.send_key('lshift', down_time=0.32)
-        self.sleep(0.57)
-        self.send_key_up('d')
-        self.sleep(0.20)
-        self.send_key_down('w')
-        self.sleep(0.40)
-        self.send_key('lshift', down_time=0.25)
-        self.sleep(0.60)
-        self.send_key_up('w')
-        self.sleep(0.15)
-        self.send_key('f', down_time=0.12)
-        self.sleep(5.00)
-        self.switch_to_avoider(check_switched=True) # 切到狗哥潜行避免碰到怪改变路径
-        self.sleep(0.10)
-        self.send_key_down('w')
-        self.sleep(0.15)
-        self.send_key_down('lshift')
-        self.sleep(0.24)
-        self.send_key('d', down_time= 0.30)
-        self.sleep(1.28)
-        self.send_key_up('lshift')
-        self.sleep(0.64)
-        self.send_key('d', down_time=0.14)
-        self.sleep(0.32)
-        self.send_key('a', down_time=0.26)
-        self.sleep(1.14)
-        self.send_key_up('w')
-        self.sleep(0.10)
-        self.switch_to_fighter(check_switched=True) # 切到早雾控怪
-        self.sleep(0.10)
-        self.send_key('a', down_time=0.24)
-        self.sleep(0.10)
-        self.send_key('w', down_time=0.32)
-        self.wait_and_interact(direction="w", is_lock=True, time_out=6.4)
-        try:
-            if self.wait_and_interact(direction=" ", is_lock=False, time_out=1.0):
-                self.send_key('s', down_time=0.12)
-                self.sleep(0.20)
-                self.send_key('e', down_time=2.42)
-                self.send_key('w', down_time=0.32)
-                self.wait_and_interact(direction="w", is_lock=True)
-                pass
-        except Exception as e:
-            pass
-        self.switch_to_avoider(check_switched=True) # 切到狗哥潜行避免碰到怪改变路径
-        self.sleep(0.15)
-        self.send_key_down('w')
-        self.sleep(0.90)
-        self.send_key('a', down_time=0.90)
-        self.sleep(0.90)
-        self.send_key_up('w')
-        self.sleep(0.10)
-        self.send_key('a', down_time=3.80)
-        self.sleep(0.10)
-        self.send_key('s', down_time=1.20)
-        self.sleep(0.10)
-        self.send_key('d', down_time=2.00)
-        self.sleep(0.10)
-        self.send_key('s', down_time=0.06)
-        self.sleep(0.10)
-        self.send_key('s', down_time=0.06)
-        self.sleep(0.10) 
-        self.click(0.50, 0.50, key="middle", down_time=0.15)
-        self.sleep(0.20)
-        self.send_key_down('w')
-        self.sleep(0.20)
-        self.send_key('lshift', down_time=1.25)
-        self.sleep(0.85)
-        self.send_key('d', down_time=0.40)
-        self.sleep(0.35)
-        self.send_key('d', down_time=0.22)
-        self.sleep(0.35)
-        self.send_key('d', down_time=0.30)
-        self.sleep(0.64)
-        self.send_key_up('w')
-        self.sleep(0.30)
-        self.switch_to_fighter(check_switched=True) # 切到早雾控怪
-        self.sleep(0.20)
-        self.send_key('s', down_time=0.10)
-        self.sleep(0.20)
-        self.send_key('e', down_time=2.60)
-        self.sleep(0.20)
-        self.send_key('f', down_time=0.12)
-        self.sleep(0.10)
-        self.send_key('f', down_time=0.12)
-        self.sleep(8.00)
-        self.send_key('w', down_time=1.25)
-        self.sleep(0.30)
-        self.send_key('f', down_time=0.12)
-        self.sleep(0.10)
-        self.send_key('f', down_time=0.12)
-        self.sleep(0.10)
