+++ src/heist_path/LobbyGotoLG1Path.py
@@ -1,0 +1,269 @@
+import time
+import re
+from src.heist_path.HeistPath import HeistPath
+
+class LobbyGotoLG1Path(HeistPath):
+    def run_path(self):
+        self.newTestLobbyGotoLG1()
+        # self.goto_lg1()
+
+    
+    # 大厅到LG1，修改版
+    def newTestLobbyGotoLG1(self):
+        self.log_round_info("寻路到LG1")
+        self.sleep(0.30)
+        self.switch_to_runner(check_switched=True)
+        self.sleep(0.10)
+        self.send_key_down('w')
+        self.sleep(0.42)
+        self.send_key('lshift', down_time=0.25)
+        self.sleep(0.42)
+        self.send_key('lshift', down_time=0.25)
+        self.sleep(0.42)
+        self.send_key('lshift', down_time=0.25)
+        self.sleep(0.42)
+        self.send_key('lshift', down_time=0.25)
+        self.sleep(0.42)
+        self.send_key_up('w')
+        self.sleep(0.20)
+        self.send_key_down('d')
+        self.sleep(0.57)
+        self.send_key('lshift', down_time=0.32)
+        self.sleep(0.57)
+        self.send_key_up('d')
+        self.sleep(0.20)
+        self.send_key_down('w')
+        self.sleep(0.40)
+        self.send_key('lshift', down_time=0.25)
+        self.sleep(0.60)
+        self.send_key_up('w')
+        self.sleep(0.15)
+        self.send_key('f', down_time=0.12)
+        self.sleep(5.00)
+        self.switch_to_avoider(check_switched=True) # 切到狗哥潜行避免碰到怪改变路径
+        self.sleep(0.10)
+        self.send_key_down('w')
+        self.sleep(0.15)
+        self.send_key_down('lshift')
+        self.sleep(0.24)
+        self.send_key('d', down_time= 0.30)
+        self.sleep(1.28)
+        self.send_key_up('lshift')
+        self.sleep(0.64)
+        self.send_key('d', down_time=0.14)
+        self.sleep(0.32)
+        self.send_key('a', down_time=0.26)
+        self.sleep(1.14)
+        self.send_key_up('w')
+        self.sleep(0.10)
+        self.switch_to_fighter(check_switched=True) # 切到早雾控怪
+        self.sleep(0.10)
+        self.send_key('a', down_time=0.24)
+        self.sleep(0.10)
+        self.send_key('w', down_time=0.32)
+        self.wait_and_interact(direction="w", is_lock=True, time_out=6.4)
+        try:
+            if self.wait_and_interact(direction=" ", is_lock=False, time_out=1.0):
+                self.send_key('s', down_time=0.12)
+                self.sleep(0.20)
+                self.send_key('e', down_time=2.42)
+                self.send_key('w', down_time=0.32)
+                self.wait_and_interact(direction="w", is_lock=True)
+                pass
+        except Exception as e:
+            pass
+        self.switch_to_avoider(check_switched=True) # 切到狗哥潜行避免碰到怪改变路径
+        self.sleep(0.15)
+        self.send_key_down('w')
+        self.sleep(0.90)
+        self.send_key('a', down_time=0.90)
+        self.sleep(0.90)
+        self.send_key_up('w')
+        self.sleep(0.10)
+        self.send_key('a', down_time=3.80)
+        self.sleep(0.10)
+        self.send_key('s', down_time=1.20)
+        self.sleep(0.10)
+        self.send_key('d', down_time=2.00)
+        self.sleep(0.10)
+        self.send_key('s', down_time=0.06)
+        self.sleep(0.10)
+        self.send_key('s', down_time=0.06)
+        self.sleep(0.10) 
+        self.click(0.50, 0.50, key="middle", down_time=0.15)
+        self.sleep(0.20)
+        self.send_key_down('w')
+        self.sleep(0.20)
+        self.send_key('lshift', down_time=1.25)
+        self.sleep(0.85)
+        self.send_key('d', down_time=0.40)
+        self.sleep(0.35)
+        self.send_key('d', down_time=0.22)
+        self.sleep(0.35)
+        self.send_key('d', down_time=0.30)
+        self.sleep(0.64)
+        self.send_key_up('w')
+        self.sleep(0.30)
+        self.switch_to_fighter(check_switched=True) # 切到早雾控怪
+        self.sleep(0.20)
+        self.send_key('s', down_time=0.10)
+        self.sleep(0.20)
+        self.send_key('e', down_time=2.60)
+        self.sleep(0.20)
+        self.send_key('f', down_time=0.12)
+        self.sleep(0.10)
+        self.send_key('f', down_time=0.12)
+        self.sleep(8.00)
+        self.send_key('w', down_time=1.25)
+        self.sleep(0.30)
+        self.send_key('f', down_time=0.12)
+        self.sleep(0.10)
+        self.send_key('f', down_time=0.12)
+        self.sleep(0.10)
+
+
+    # 旧路径参考自 B站 UP 主：@早柚大魔王
+    def goto_lg1(self):
+        self.log_round_info("寻路到LG1")
+        self.switch_to_runner(check_switched=True)
+        self.sleep(0.81)
+        self.send_key_down("w")
+        self.sleep(0.32)
+        self.send_key_down("lshift")
+        self.sleep(0.16)
+        self.send_key_up("lshift")
+        self.sleep(2.68)
+        self.send_key_down("d")
+        self.sleep(2.55)
+        self.send_key_up("d")
+        self.sleep(0.37)
+        self.wait_and_interact(direction="w", is_lock=True)
+        self.sleep(0.5)
+        self.send_key_down("w")
+        self.sleep(1.8)
+
+        deadline = time.time() + 5.5
+        while time.time() < deadline:
+            self.send_key_down("space")
+            self.sleep(0.06)
+            self.send_key_up("space")
+            self.sleep(0.26)
+
+        if not self.wait_and_interact(direction="w", key_up_sleep=0, is_lock=True):
+            self.goto_lg1_interrupted()
+
+        self.sleep(0.5)
+        self.send_key_down("w")
+        self.sleep(0.2)
+        self.send_key("lshift")
+        self.sleep(0.2)
+        self.send_key_down("d")
+        self.sleep(0.75)
+        self.send_key_up("d")
+        self.sleep(0.01)
+        self.send_key_down("a")
+        self.sleep(1)
+        self.send_key_up("a")
+        self.sleep(0.5)
+        self.send_key_up("w")
+        self.sleep(0.29)
+        self.send_key_down("a")
+        self.sleep(0.29)
+        self.send_key("lshift")
+        self.sleep(3.20)
+        self.send_key_up("a")
+        self.sleep(0.04)
+        self.send_key_down("s")
+        self.sleep(0.29)
+        self.send_key("lshift")
+        self.sleep(1.50)
+        self.send_key_up("s")
+        self.sleep(0.04)
+        self.send_key_down("d")
+        self.sleep(0.29)
+        self.send_key("lshift")
+        self.sleep(2.50)
+        self.send_key_up("d")
+        self.sleep(0.40)
+        self.send_key_down("a")
+        self.sleep(0.71)
+        self.send_key_up("a")
+        self.sleep(0.36)
+        self.send_key_down("s")
+        self.sleep(1.50)
+        self.send_key_up("s")
+        self.sleep(0.14)
+        self.send_key_down("f") # start pick
+        self.sleep(0.04)
+        self.send_key_down("w")
+        self.sleep(2.5)
+        self.send_key_up("w")
+        self.sleep(0.10)
+        self.send_key_up("f") # end pick
+        self.sleep(0.13)
+        self.send_key_down("s")
+        self.sleep(0.14)
+        self.send_key_up("s")
+        self.sleep(0.20)
+        self.clear_current_combat()
+        self.send_key_down("f")
+        self.sleep(0.5)
+        self.send_key_down("w")
+        self.sleep(0.22)
+        self.send_key_down("lshift")
+        self.sleep(0.10)
+        self.send_key_up("lshift")
+        self.sleep(1)
+        self.send_key_up("f")
+        self.sleep(0.1)
+        self.send_key_down("d")
+        self.sleep(1)
+        self.send_key_up("w")
+        self.sleep(1.5)
+        self.send_key_up("d")
+        self.sleep(0.35)
+        self.send_key_down("a")
+        self.sleep(0.88)
+        self.send_key_up("a")
+        self.sleep(0.30)
+        self.send_key_down("s")
+        self.sleep(0.43)
+        self.send_key_down("lshift")
+        self.sleep(0.13)
+        self.send_key_up("lshift")
+        self.sleep(1.4)
+        self.send_key_down("a")
+        self.sleep(0.53)
+        self.send_key_up("a")
+        self.sleep(1.64)
+        self.wait_and_interact(direction="s")
+        self.sleep(0.50)
+        self.send_key_down("s")
+        self.sleep(0.10)
+        self.wait_and_interact(direction="s")
+
+    def goto_lg1_interrupted(self):
+        self.clear_current_combat()
+        self.send_key_down("w")
+        self.sleep(2.02)
+        self.send_key_up("w")
+        self.sleep(0.51)
+        self.send_key_down("s")
+        self.sleep(0.60)
+        self.send_key_up("s")
+        self.sleep(0.22)
+        self.send_key_down("d")
+        self.sleep(3.41)
+        self.send_key_up("d")
+        self.sleep(0.32)
+        self.send_key_down("a")
+        self.sleep(1.16)
+        self.send_key_up("a")
+        self.sleep(0.20)
+        self.send_key_down("w")
+        self.sleep(1.51)
+        self.send_key_up("w")
+        self.sleep(0.11)
+        self.wait_and_interact(direction="w", is_lock=True)
+        self.sleep(0.5)
+
